Move functions from leveler
